+++ scripts/test/TemplateGenerator/Template.t
@@ -32,6 +32,19 @@ $Kernel::OM->ObjectParamAdd(
 );
 my $Helper = $Kernel::OM->Get('Kernel::System::UnitTest::Helper');
 
+# Do not check email addresses.
+$Helper->ConfigSettingChange(
+    Key   => 'CheckEmailAddresses',
+    Value => 0,
+);
+
+# Do not check RichText.
+$Helper->ConfigSettingChange(
+    Valid => 1,
+    Key   => 'Frontend::RichText',
+    Value => 0,
+);
+
 my $TicketObject = $Kernel::OM->Get('Kernel::System::Ticket');
 
 # Create test ticket.
@@ -50,110 +63,321 @@ my $TicketID     = $TicketObject->TicketCreate(
 );
 $Self->True(
     $TicketID,
-    "Ticket is created - TicketID $TicketID",
+    "TicketID $TicketID is created",
 );
 
 my $ArticleBackendObject = $Kernel::OM->Get('Kernel::System::Ticket::Article')->BackendForChannel(
     ChannelName => 'Phone',
 );
 
-my $ArticleID = $ArticleBackendObject->ArticleCreate(
-    TicketID             => $TicketID,
-    IsVisibleForCustomer => 0,
-    SenderType           => 'agent',
-    From                 => 'Some Agent <otrs@example.com>',
-    To                   => 'Suplier<suplier@example.com>',
-    Subject              => 'Email for suplier',
-    Body                 => "Line1\nLine2\nLine3",
-    Charset              => 'utf8',
-    MimeType             => 'text/plain',
-    HistoryType          => 'OwnerUpdate',
-    HistoryComment       => 'Some free text!',
-    UserID               => 1,
+my $LastAgentSubject      = 'Article#3-agent';
+my $LastAgentSubject9     = 'Article#3 [...]';
+my $LastAgentBody         = "agent-Article#3-Line1\nagent-Article#3-Line2\nagentArticle#3-Line3";
+my $LastAgentBody2        = "> agent-Article#3-Line1\n> agent-Article#3-Line2";
+my $LastCustomerSubject   = 'Article#6-customer';
+my $LastCustomerSubject12 = 'Article#6-cu [...]';
+my $LastCustomerBody      = "customer-Article#6-Line1\ncustomer-Article#6-Line2\ncustomerArticle#6-Line3";
+my $LastCustomerBody1     = "> customer-Article#6-Line1";
+
+my @Configs = (
+    {
+        SenderType => 'agent',
+        Subject    => 'Article#1-agent',
+        Body       => "agent-Article#1-Line1\nagent-Article#1-Line2\nagentArticle#1-Line3",
+    },
+    {
+        SenderType => 'agent',
+        Subject    => 'Article#2-agent',
+        Body       => "agent-Article#2-Line1\nagent-Article#2-Line2\nagentArticle#2-Line3",
+    },
+    {
+        SenderType => 'agent',
+        Subject    => $LastAgentSubject,
+        Body       => $LastAgentBody,
+    },
+    {
+        SenderType => 'customer',
+        Subject    => 'Article#4-customer',
+        Body       => "customer-Article#4-Line1\ncustomer-Article#4-Line2\ncustomerArticle#4-Line3",
+    },
+    {
+        SenderType => 'customer',
+        Subject    => 'Article#5-customer',
+        Body       => "customer-Article#5-Line1\ncustomer-Article#5-Line2\ncustomerArticle#5-Line3",
+    },
+    {
+        SenderType => 'customer',
+        Subject    => $LastCustomerSubject,
+        Body       => $LastCustomerBody,
+    },
 );
-$Self->True(
-    $ArticleID,
-    "ArticleID $ArticleID is created"
-);
+
+my @Articles;
+
+for my $Config (@Configs) {
+    my $ArticleID = $ArticleBackendObject->ArticleCreate(
+        %{$Config},
+        TicketID             => $TicketID,
+        IsVisibleForCustomer => 0,
+        From                 => 'Some Agent <otrs@example.com>',
+        To                   => 'Suplier<suplier@example.com>',
+        Charset              => 'utf8',
+        MimeType             => 'text/plain',
+        HistoryType          => 'OwnerUpdate',
+        HistoryComment       => 'Some free text!',
+        UserID               => 1,
+    );
+    $Self->True(
+        $ArticleID,
+        "ArticleID $ArticleID is created"
+    );
+    my %ArticleData = $ArticleBackendObject->ArticleGet(
+        TicketID  => $TicketID,
+        ArticleID => $ArticleID,
+    );
+    push @Articles, \%ArticleData;
+}
 
 # Get ticket and article data for tests.
 my %TicketData = $Kernel::OM->Get('Kernel::System::Ticket')->TicketGet(
     TicketID      => $TicketID,
     DynamicFields => 1,
 );
-my %ArticleData = $ArticleBackendObject->ArticleGet(
-    TicketID  => $TicketID,
-    ArticleID => $ArticleID,
+
+# Define for which template types certain tags are supported.
+my %Supported = (
+    Answer  => 1,
+    Forward => 1,
+    Note    => 1,
 );
-my %Data = ( %TicketData, %ArticleData );
 
 my @Tests = (
     {
-        Name           => 'Test supported tag - <OTRS_CONFIG_ScriptAlias>',
+        Name           => 'Supported tag - <OTRS_CONFIG_ScriptAlias>',
         TemplateText   => 'Thank you for your email. <OTRS_CONFIG_ScriptAlias>',
         ExpectedResult => 'Thank you for your email. ' . $ConfigObject->Get('ScriptAlias'),
     },
     {
-        Name           => 'Test unsupported tags',
-        TemplateText   => 'Test: <OTRS_AGENT_SUBJECT> <OTRS_AGENT_BODY> <OTRS_CUSTOMER_BODY> <OTRS_CUSTOMER_SUBJECT>',
-        ExpectedResult => 'Test: - - - -',
-    },
-    {
-        Name => 'Test unsupported tags with [n]',
-        TemplateText =>
-            'Test: <OTRS_AGENT_SUBJECT[2]> <OTRS_AGENT_BODY[2]> <OTRS_CUSTOMER_BODY[2]> <OTRS_CUSTOMER_SUBJECT[2]>',
-        ExpectedResult => 'Test: - - - -',
-        TicketID       => $TicketID,
-        Data           => \%Data,
-    },
-    {
-        Name => 'Test supported tags - <OTRS_TICKET_*> without TicketID',
+        Name => 'Supported tags - <OTRS_TICKET_*> without TicketID',
         TemplateText =>
             'Options of the ticket data (e. g. <OTRS_TICKET_TicketNumber>, <OTRS_TICKET_TicketID>, <OTRS_TICKET_Queue>)',
         ExpectedResult => 'Options of the ticket data (e. g. -, -, -)',
     },
     {
-        Name => 'Test supported tags - <OTRS_TICKET_*>  with TicketID',
+        Name => 'Supported tags - <OTRS_TICKET_*>  with TicketID',
         TemplateText =>
             'Options of the ticket data (e. g. <OTRS_TICKET_TicketNumber>, <OTRS_TICKET_TicketID>, <OTRS_TICKET_Queue>, <OTRS_TICKET_State>)',
         ExpectedResult => "Options of the ticket data (e. g. $TicketNumber, $TicketID, Raw, open)",
         TicketID       => $TicketID,
+    },
+    {
+        Name           => 'Tag <OTRS_AGENT_SUBJECT>',
+        TemplateText   => 'Test: <OTRS_AGENT_SUBJECT>',
+        TicketID       => $TicketID,
+        TemplateResult => {
+            Note      => "Test: $LastAgentSubject",
+            Supported => {
+                $Articles[0]->{ArticleID} => "Test: $Articles[0]->{Subject}",
+                $Articles[1]->{ArticleID} => "Test: $Articles[1]->{Subject}",
+                $Articles[2]->{ArticleID} => "Test: $Articles[2]->{Subject}",
+                $Articles[3]->{ArticleID} => 'Test: -',
+                $Articles[4]->{ArticleID} => 'Test: -',
+                $Articles[5]->{ArticleID} => 'Test: -',
+            },
+            Unsupported => 'Test: -',
+        }
+    },
+    {
+        Name           => 'Tag <OTRS_AGENT_SUBJECT[9]>',
+        TemplateText   => 'Test: <OTRS_AGENT_SUBJECT[9]>',
+        TicketID       => $TicketID,
+        TemplateResult => {
+            Note      => "Test: $LastAgentSubject9",
+            Supported => {
+                $Articles[0]->{ArticleID} => 'Test: Article#1 [...]',
+                $Articles[1]->{ArticleID} => 'Test: Article#2 [...]',
+                $Articles[2]->{ArticleID} => 'Test: Article#3 [...]',
+                $Articles[3]->{ArticleID} => 'Test: -',
+                $Articles[4]->{ArticleID} => 'Test: -',
+                $Articles[5]->{ArticleID} => 'Test: -',
+            },
+            Unsupported => 'Test: -',
+        }
+    },
+    {
+        Name           => 'Tag <OTRS_AGENT_BODY>',
+        TemplateText   => 'Test: <OTRS_AGENT_BODY>',
+        TicketID       => $TicketID,
+        TemplateResult => {
+            Note      => "Test: $LastAgentBody",
+            Supported => {
+                $Articles[0]->{ArticleID} => "Test: $Articles[0]->{Body}",
+                $Articles[1]->{ArticleID} => "Test: $Articles[1]->{Body}",
+                $Articles[2]->{ArticleID} => "Test: $Articles[2]->{Body}",
+                $Articles[3]->{ArticleID} => 'Test: -',
+                $Articles[4]->{ArticleID} => 'Test: -',
+                $Articles[5]->{ArticleID} => 'Test: -',
+            },
+            Unsupported => 'Test: -',
+        }
+    },
+    {
+        Name           => 'Tag <OTRS_AGENT_BODY[2]>',
+        TemplateText   => 'Test: <OTRS_AGENT_BODY[2]>',
+        TicketID       => $TicketID,
+        TemplateResult => {
+            Note      => "Test: $LastAgentBody2",
+            Supported => {
+                $Articles[0]->{ArticleID} => "Test: > agent-Article#1-Line1\n> agent-Article#1-Line2",
+                $Articles[1]->{ArticleID} => "Test: > agent-Article#2-Line1\n> agent-Article#2-Line2",
+                $Articles[2]->{ArticleID} => "Test: > agent-Article#3-Line1\n> agent-Article#3-Line2",
+                $Articles[3]->{ArticleID} => 'Test: -',
+                $Articles[4]->{ArticleID} => 'Test: -',
+                $Articles[5]->{ArticleID} => 'Test: -',
+            },
+            Unsupported => 'Test: -',
+        }
+    },
+    {
+        Name           => 'Tag <OTRS_CUSTOMER_SUBJECT>',
+        TemplateText   => 'Test: <OTRS_CUSTOMER_SUBJECT>',
+        TicketID       => $TicketID,
+        TemplateResult => {
+            Note      => "Test: $LastCustomerSubject",
+            Supported => {
+                $Articles[0]->{ArticleID} => "Test: $Articles[0]->{Subject}",
+                $Articles[1]->{ArticleID} => "Test: $Articles[1]->{Subject}",
+                $Articles[2]->{ArticleID} => "Test: $Articles[2]->{Subject}",
+                $Articles[3]->{ArticleID} => "Test: $Articles[3]->{Subject}",
+                $Articles[4]->{ArticleID} => "Test: $Articles[4]->{Subject}",
+                $Articles[5]->{ArticleID} => "Test: $Articles[5]->{Subject}",
+            },
+            Unsupported => 'Test: -',
+        }
+    },
+    {
+        Name           => 'Tag <OTRS_CUSTOMER_SUBJECT[12]>',
+        TemplateText   => 'Test: <OTRS_CUSTOMER_SUBJECT[12]>',
+        TicketID       => $TicketID,
+        TemplateResult => {
+            Note      => "Test: $LastCustomerSubject12",
+            Supported => {
+                $Articles[0]->{ArticleID} => 'Test: Article#1-ag [...]',
+                $Articles[1]->{ArticleID} => 'Test: Article#2-ag [...]',
+                $Articles[2]->{ArticleID} => 'Test: Article#3-ag [...]',
+                $Articles[3]->{ArticleID} => 'Test: Article#4-cu [...]',
+                $Articles[4]->{ArticleID} => 'Test: Article#5-cu [...]',
+                $Articles[5]->{ArticleID} => 'Test: Article#6-cu [...]',
+            },
+            Unsupported => 'Test: -',
+        }
+    },
+    {
+        Name           => 'Tag <OTRS_CUSTOMER_BODY>',
+        TemplateText   => 'Test: <OTRS_CUSTOMER_BODY>',
+        TicketID       => $TicketID,
+        TemplateResult => {
+            Note      => "Test: $LastCustomerBody",
+            Supported => {
+                $Articles[0]->{ArticleID} => "Test: $Articles[0]->{Body}",
+                $Articles[1]->{ArticleID} => "Test: $Articles[1]->{Body}",
+                $Articles[2]->{ArticleID} => "Test: $Articles[2]->{Body}",
+                $Articles[3]->{ArticleID} => "Test: $Articles[3]->{Body}",
+                $Articles[4]->{ArticleID} => "Test: $Articles[4]->{Body}",
+                $Articles[5]->{ArticleID} => "Test: $Articles[5]->{Body}",
+            },
+            Unsupported => 'Test: -',
+        }
+    },
+    {
+        Name           => 'Tag <OTRS_CUSTOMER_BODY[1]>',
+        TemplateText   => 'Test: <OTRS_CUSTOMER_BODY[1]>',
+        TicketID       => $TicketID,
+        TemplateResult => {
+            Note      => "Test: $LastCustomerBody1",
+            Supported => {
+                $Articles[0]->{ArticleID} => "Test: > agent-Article#1-Line1",
+                $Articles[1]->{ArticleID} => "Test: > agent-Article#2-Line1",
+                $Articles[2]->{ArticleID} => "Test: > agent-Article#3-Line1",
+                $Articles[3]->{ArticleID} => "Test: > customer-Article#4-Line1",
+                $Articles[4]->{ArticleID} => "Test: > customer-Article#5-Line1",
+                $Articles[5]->{ArticleID} => "Test: > customer-Article#6-Line1",
+            },
+            Unsupported => 'Test: -',
+        }
     },
 );
 
 my $StandardTemplateObject  = $Kernel::OM->Get('Kernel::System::StandardTemplate');
 my $TemplateGeneratorObject = $Kernel::OM->Get('Kernel::System::TemplateGenerator');
 
+TEST:
 for my $Test (@Tests) {
+    for my $TemplateType (qw(Answer Forward Create Note Email PhoneCall)) {
 
-    # Create standard template.
-    my $TemplateID = $StandardTemplateObject->StandardTemplateAdd(
-        Name         => $Helper->GetRandomID() . '-StandardTemplate',
-        Template     => $Test->{TemplateText},
-        ContentType  => 'text/plain; charset=utf-8',
-        TemplateType => 'Answer',
-        ValidID      => 1,
-        UserID       => 1,
-    );
-    $Self->True(
-        $TemplateID,
-        "StandardTemplate is created - ID $TemplateID",
-    );
+        # Create standard template.
+        my $TemplateID = $StandardTemplateObject->StandardTemplateAdd(
+            Name         => $Helper->GetRandomID() . '-StandardTemplate',
+            Template     => $Test->{TemplateText},
+            ContentType  => 'text/plain; charset=utf-8',
+            TemplateType => $TemplateType,
+            ValidID      => 1,
+            UserID       => 1,
+        );
+        $Self->True(
+            $TemplateID,
+            "'$TemplateType' type - TemplateID $TemplateID is created",
+        );
 
-    my $Template = $TemplateGeneratorObject->Template(
-        TemplateID => $TemplateID,
-        TicketID   => $Test->{TicketID} // '',
-        Data       => $Test->{Data} // {},
-        UserID     => 1,
-    );
+        # Check template text.
+        if ( $Test->{ExpectedResult} ) {
+            my $Template = $TemplateGeneratorObject->Template(
+                TemplateID => $TemplateID,
+                TicketID   => $Test->{TicketID},
+                Data       => {},
+                UserID     => 1,
+            );
+            $Self->Is(
+                $Template,
+                $Test->{ExpectedResult},
+                "'$TemplateType' type - $Test->{Name}",
+            );
+        }
+        elsif ( $Test->{TemplateResult} ) {
 
-    # Check template text.
-    $Self->Is(
-        $Template,
-        $Test->{ExpectedResult},
-        $Test->{Name},
-    );
+            # Test for all agent and customer articles.
+            for my $Article (@Articles) {
+                my $Template = $TemplateGeneratorObject->Template(
+                    TemplateID => $TemplateID,
+                    TicketID   => $Test->{TicketID},
+                    Data       => { %TicketData, %{$Article} },
+                    UserID     => 1,
+                );
 
+                if ( $Supported{$TemplateType} ) {
+                    my $ExpectedResult = $Test->{TemplateResult}->{Supported}->{ $Article->{ArticleID} } // '';
+
+                    # For Note template, there is last article data.
+                    if ( $TemplateType eq 'Note' ) {
+                        $ExpectedResult = $Test->{TemplateResult}->{Note};
+                    }
+
+                    $Self->Is(
+                        $Template,
+                        $ExpectedResult,
+                        "'$TemplateType' type - $Article->{Subject} - $Test->{Name}",
+                    );
+                }
+                else {
+                    $Self->Is(
+                        $Template,
+                        $Test->{TemplateResult}->{Unsupported},
+                        "'$TemplateType' type - $Article->{Subject} - $Test->{Name}",
+                    );
+                }
+            }
+        }
+    }
 }
 
 # Cleanup is done by RestoreDatabase.